pyproject.toml
[project]
name = "law-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.3.5",
    "openai>=2.9.0",
    "sentence-transformers>=5.1.2",
    "tiktoken>=0.12.0",
    "torchvision>=0.24.1",
    "unsloth[colab-new]",
    "unstructured>=0.18.21",
]

[tool.uv.sources]
unsloth = { git = "https://github.com/unslothai/unsloth.git" }
